fix: CLI review — backend validation, deduplication, cleanup
- Fix: follow --payload/--http now rejects invalid --backend values
  instead of silently falling through to tshark (input validation bug)
- Fix: native-only commands (detail, issues, audit, follow) now parse
  BackendKind first, so unknown backends get "unsupported backend: X"
  instead of the misleading "does not support" message
- Deduplicate summary line format string into summary::format_line()
  used by summary.rs (native + tshark) and follow.rs
- Deduplicate stats rendering into CaptureStats struct with
  render_json()/render_human() methods, eliminating ~60% shared code
  between run_native and run_tshark
- Remove timestamp.rs one-line re-export; use fireshark_core::format_utc
  directly in all modules
- Use DecodeIssueKind::Display in detail.rs and issues.rs instead of
  manual match arms; update Display to produce "Checksum mismatch"
- Fix diff.rs isize overflow: use i64 for delta calculations
- Remove redundant serde_json dev-dependency

+++ crates/fireshark-cli/src/diff.rs
@@ -105,8 +105,8 @@ pub fn run(
         );
         println!();
 
-        let packet_delta = result.b_packet_count as isize - result.a_packet_count as isize;
-        let stream_delta = result.b_stream_count as isize - result.a_stream_count as isize;
+        let packet_delta = result.b_packet_count as i64 - result.a_packet_count as i64;
+        let stream_delta = result.b_stream_count as i64 - result.a_stream_count as i64;
         println!("  Packet delta:   {:+}", packet_delta);
         println!("  Stream delta:   {:+}", stream_delta);
